consolidate conn*.go files after dropping old Go support
